Abstract player entity to humanoid entity
